+++ client/src/api/block_builder/input_selection/core/requirement/issuer.rs
@@ -1,10 +1,9 @@
 // Copyright 2023 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use super::{InputSelection, Requirement};
+use super::{Error, InputSelection, Requirement};
 use crate::{
     block::{address::Address, output::AliasTransition},
-    error::{Error, Result},
     secret::types::InputSigningData,
 };
 
@@ -14,7 +13,7 @@ impl InputSelection {
     pub(crate) fn fulfill_issuer_requirement(
         &mut self,
         address: Address,
-    ) -> Result<Vec<(InputSigningData, Option<AliasTransition>)>> {
+    ) -> Result<Vec<(InputSigningData, Option<AliasTransition>)>, Error> {
         log::debug!("Treating {address:?} issuer requirement as a sender requirement");
 
         match self.fulfill_sender_requirement(address) {