+++ client/src/api/block_builder/input_selection/core/requirement/nft.rs
@@ -1,10 +1,9 @@
 // Copyright 2023 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use super::{InputSelection, Requirement};
+use super::{Error, InputSelection, Requirement};
 use crate::{
     block::output::{AliasTransition, NftId, Output, OutputId},
-    error::{Error, Result},
     secret::types::InputSigningData,
 };
 
@@ -35,7 +34,7 @@ impl InputSelection {
     pub(crate) fn fulfill_nft_requirement(
         &mut self,
         nft_id: NftId,
-    ) -> Result<Vec<(InputSigningData, Option<AliasTransition>)>> {
+    ) -> Result<Vec<(InputSigningData, Option<AliasTransition>)>, Error> {
         // Check if the requirement is already fulfilled.
         if let Some(input) = self
             .selected_inputs