+++ client/tests/input_selection/foundry_outputs.rs
@@ -4,13 +4,12 @@
 use std::{collections::HashSet, str::FromStr};
 
 use iota_client::{
-    api::input_selection::{Burn, InputSelection, Requirement},
+    api::input_selection::{Burn, Error, InputSelection, Requirement},
     block::{
         address::{Address, AliasAddress},
         output::{AliasId, AliasOutputBuilder, AliasTransition, FoundryId, Output, SimpleTokenScheme, TokenId},
         protocol::protocol_parameters,
     },
-    Error,
 };
 use primitive_types::U256;