+++ client/CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `high_level` module moved to `api`;
 - Added `#[serde(rename_all = "camelCase")]` to `NodeAuth`;
 - `Topic::try_new` parameter from `String` to `impl Into<String>`;
+- All `Error` variant related to ISA have been moved to a new ISA `Error` enum;
 
 ### Removed